Re-factor test_sgrid.py
